feat(site): mount the configurator inside the site layout
