+++ configurator/index.html
@@ -7,15 +7,11 @@
     <link rel="icon" href="/favicon.ico" />
   </head>
   <body>
-    <header>
+    <header class="standalone-header">
       <h1><a href="/">HomeRacker</a> Configurator</h1>
       <p>Plan a rack, preview it, and get the parts list to print.</p>
     </header>
-    <div id="layout">
-      <aside id="controls"></aside>
-      <main><canvas id="viewer"></canvas></main>
-      <section id="bom"></section>
-    </div>
+    <div id="app"></div>
     <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
