+++ configurator/src/configurator.css
@@ -1,0 +1,235 @@
+/*
+ * Configurator component styles. Reads the site's design tokens when mounted inside
+ * homeracker.org and falls back to matching values when running standalone.
+ */
+
+.cfg {
+  --cfg-bg: var(--bg, #0b0c10);
+  --cfg-panel: var(--panel, #15171c);
+  --cfg-panel-2: var(--panel-2, #1c1f26);
+  --cfg-line: var(--line, #2a2e38);
+  --cfg-text: var(--text, #e8e6df);
+  --cfg-muted: var(--muted, #a3a49f);
+  --cfg-accent: var(--accent, #f7b600);
+  --cfg-accent-ink: var(--accent-ink, #1a1300);
+  --cfg-danger: var(--danger, #ff4d6a);
+  --cfg-label: var(--font-label, "Barlow Semi Condensed", "Arial Narrow", sans-serif);
+  --cfg-display: var(--font-display, "Barlow Condensed", "Arial Narrow", sans-serif);
+
+  display: grid;
+  grid-template-columns: 19rem minmax(0, 1fr) 23rem;
+  min-height: 640px;
+  height: calc(100vh - 4.5rem);
+  color: var(--cfg-text);
+  border-top: 1px solid var(--cfg-line);
+  border-bottom: 1px solid var(--cfg-line);
+}
+
+.cfg-controls,
+.cfg-bom {
+  padding: 1.25rem;
+  overflow: auto;
+  background: var(--cfg-panel);
+}
+
+.cfg-controls {
+  border-right: 1px solid var(--cfg-line);
+}
+
+.cfg-bom {
+  border-left: 1px solid var(--cfg-line);
+}
+
+.cfg-stage {
+  position: relative;
+  min-height: 60vh;
+  background: var(--cfg-bg);
+}
+
+.cfg-canvas {
+  display: block;
+  width: 100%;
+  height: 100%;
+}
+
+.cfg fieldset {
+  border: 1px solid var(--cfg-line);
+  padding: 0.6rem 0.9rem 0.9rem;
+  margin: 0 0 1rem;
+}
+
+.cfg legend {
+  font-family: var(--cfg-label);
+  font-weight: 600;
+  font-size: 0.85rem;
+  letter-spacing: 0.1em;
+  text-transform: uppercase;
+  color: var(--cfg-accent);
+  padding: 0 0.4rem;
+}
+
+.cfg .field {
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+  margin: 0.55rem 0;
+}
+
+.cfg .field.inline {
+  flex-direction: row;
+  align-items: center;
+  gap: 0.6rem;
+  min-height: 44px;
+}
+
+.cfg label {
+  font-size: 0.95rem;
+  color: var(--cfg-text);
+}
+
+.cfg input[type="number"],
+.cfg input[type="text"],
+.cfg select {
+  font: inherit;
+  min-height: 44px;
+  padding: 0.4rem 0.6rem;
+  color: var(--cfg-text);
+  background: var(--cfg-bg);
+  border: 1px solid var(--cfg-line);
+  border-radius: 0;
+}
+
+.cfg input:focus-visible,
+.cfg select:focus-visible {
+  outline: 2px solid var(--cfg-accent);
+  outline-offset: 1px;
+  border-color: var(--cfg-accent);
+}
+
+.cfg input[type="checkbox"],
+.cfg input[type="radio"] {
+  width: 1.15rem;
+  height: 1.15rem;
+  accent-color: var(--cfg-accent);
+  margin: 0;
+}
+
+.cfg output,
+.cfg .readout {
+  font-family: var(--cfg-label);
+  font-size: 0.9rem;
+  color: var(--cfg-muted);
+  margin: 0;
+}
+
+.cfg #issues {
+  color: var(--cfg-danger);
+  padding-left: 1.2rem;
+  margin: 0.5rem 0;
+}
+
+.cfg #issues:empty {
+  display: none;
+}
+
+.cfg h2 {
+  font-family: var(--cfg-display);
+  font-weight: 700;
+  font-size: 1.9rem;
+  line-height: 1;
+  text-transform: uppercase;
+  margin: 0 0 0.4rem;
+}
+
+.cfg table {
+  border-collapse: collapse;
+  width: 100%;
+  margin-top: 0.75rem;
+}
+
+.cfg td,
+.cfg th {
+  text-align: left;
+  vertical-align: top;
+  padding: 0.35rem 0.4rem;
+  border-bottom: 1px solid var(--cfg-line);
+}
+
+.cfg tr.group th {
+  padding-top: 1rem;
+  font-family: var(--cfg-label);
+  font-weight: 600;
+  font-size: 0.85rem;
+  letter-spacing: 0.1em;
+  text-transform: uppercase;
+  color: var(--cfg-accent);
+  border-bottom-color: var(--cfg-accent);
+}
+
+.cfg td.qty {
+  text-align: right;
+  font-variant-numeric: tabular-nums;
+  font-weight: 600;
+  width: 3rem;
+}
+
+.cfg td.note {
+  font-size: 0.8rem;
+  color: var(--cfg-muted);
+}
+
+.cfg tr.total td {
+  font-weight: 600;
+  border-bottom: none;
+  padding-top: 0.8rem;
+}
+
+.cfg button {
+  margin-top: 1.2rem;
+  min-height: 44px;
+  padding: 0.6rem 1.2rem;
+  font-family: var(--cfg-label);
+  font-weight: 600;
+  font-size: 1rem;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+  color: var(--cfg-accent-ink);
+  background: var(--cfg-accent);
+  border: none;
+  cursor: pointer;
+  touch-action: manipulation;
+  clip-path: polygon(0 0, calc(100% - 10px) 0, 100% 10px, 100% 100%, 10px 100%, 0 calc(100% - 10px));
+}
+
+.cfg button:hover,
+.cfg button:focus-visible {
+  background: #ffcd3a;
+}
+
+.cfg textarea {
+  width: 100%;
+  min-height: 12rem;
+  margin-top: 0.5rem;
+  font: inherit;
+  color: var(--cfg-text);
+  background: var(--cfg-bg);
+  border: 1px solid var(--cfg-line);
+}
+
+@media (max-width: 1000px) {
+  .cfg {
+    grid-template-columns: minmax(0, 1fr);
+    height: auto;
+  }
+
+  .cfg-controls,
+  .cfg-bom {
+    border: none;
+    border-bottom: 1px solid var(--cfg-line);
+  }
+
+  .cfg-stage {
+    height: 60vh;
+    min-height: 360px;
+  }
+}
